# shared/openapi.yaml
openapi: 3.1.0
...
info:
  title: Syl API
  version: 1.0.0
  summary: The contract between Syl's service, her iOS app, and the web admin.
  description: |
    # Conventions

    ## Envelope

    Every response body — success or failure — is one of:

    ```json
    { "success": true,  "data": { ... } }
    { "success": false, "error": { "code": "NOT_FOUND", "message": "...", "retryable": false } }
    ```

    There is no third shape. A client that cannot parse one of these two has
    hit something that is not Syl (a proxy, a captive portal, a Tailscale
    error page) and should treat it as a transport failure, not an API error.

    ## Field naming

    **camelCase for every JSON field, with exactly one exception: `ttl_ms` on
    the `presence` frame.** That field is quoted as literal JSON in the
    character proposal and in `syl-001.2.2`, three separate sources pin the
    spelling, and silently "fixing" it is how a contract stops matching the
    thing it was derived from. It is the only snake_case identifier on the
    wire. Do not add a second one.

    Note that the service's SQLite columns are snake_case (child B). The
    mapping is at the store boundary, not on the wire.

    ## Identifiers

    Every resource id is `syl:<type>:<uuidv7>` — type-prefixed so a dangling
    reference is legible in a log line, UUIDv7 so ids sort by creation time.
    The convention is fixed before the first row is written because every
    graph edge will reference it forever.

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /reminders:
    get:
      tags: [reminders]
      operationId: listReminders
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: state
          in: query
          required: false
          schema: { $ref: "#/components/schemas/ReminderDeliveryState" }
        - name: dueBefore
          in: query
          required: false
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: A page of reminders.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/ReminderPage" }
        default: { $ref: "#/components/responses/Error" }
    post:
      tags: [reminders]
      operationId: createReminder
      description: |
        `text` is required and is composed **at creation time, in Syl's
        voice**. Delivery is a zero-turn path — read the due row, write the
        outbox row, push — so nothing downstream can improve the wording. A
        perfectly reliable pipeline delivering flat, characterless text is
        mechanically correct and tonally wrong.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReminderRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Reminder" }
        default: { $ref: "#/components/responses/Error" }
  /reminders/{reminderId}:
    get:
      tags: [reminders]
      operationId: getReminder
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      responses:
        "200":
          description: The reminder.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Reminder" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      tags: [reminders]
      operationId: updateReminder
      parameters:
        - $ref: "#/components/parameters/ReminderId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateReminderRequest" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Reminder" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [reminders]
      operationId: cancelReminder
      summary: Cancel a reminder. Rows are closed, never deleted.
      parameters:
        - $ref: "#/components/parameters/ReminderId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Cancelled. `deliveryState` becomes `cancelled`.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Reminder" }
        default: { $ref: "#/components/responses/Error" }
  /reminders/{reminderId}/complete:
    post:
      tags: [reminders]
      operationId: completeReminder
      parameters:
        - $ref: "#/components/parameters/ReminderId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Completed.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Reminder" }
        default: { $ref: "#/components/responses/Error" }
  /reminders/{reminderId}/snooze:
    post:
      tags: [reminders]
      operationId: snoozeReminder
      summary: Defer a reminder. Authority is server-side, always.
      description: |
        The notification action lives on the device; the authority does not.
        A phone that is wiped, restored, or replaced would take device-local
        deferrals with it, and a deferral that vanishes is the one outcome
        the project forbids.

        **The server must return a strictly later instant.** If the requested
        instant is not strictly after the current `nextFireAt`, this fails
        with `422 DEFERRAL_NOT_LATER` rather than silently accepting a
        deferral that would drop the reminder.
      parameters:
        - $ref: "#/components/parameters/ReminderId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SnoozeReminderRequest" }
      responses:
        "200":
          description: Deferred to a strictly later instant.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/OkEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: "#/components/schemas/Reminder" }
        default: { $ref: "#/components/responses/Error" }

  # --------------------------------------------------------------- todos ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated, opaque, unique per logical operation. Required on
        every write — the mobile outbox retries by design and will duplicate
        without it.
      schema: { type: string, minLength: 8, maxLength: 128 }
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque page cursor from the previous page's `nextCursor`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    ReminderId:
      name: reminderId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
  responses:
    Error:
      description: |
        Any failure. The HTTP status is advisory; `error.code` is the
        contract. `error.retryable` tells the client's backoff whether to try
        again — the Tailscale extension is torn down when idle, so the first
        request after a wake can fail while the tunnel re-establishes, and
        treating that as "server down" makes a healthy app feel broken.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    Id:
      type: string
      description: "`syl:<type>:<uuidv7>` — type-prefixed, time-sortable."
      pattern: "^syl:[a-z][a-z_]*:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
      examples: ["syl:reminder:0198f2c1-4a3b-7d21-9f00-1a2b3c4d5e6f"]
    Instant:
      type: string
      format: date-time
      description: RFC 3339, UTC, millisecond precision. Never a fixed offset.
      examples: ["2026-08-09T07:00:03.114Z"]
    WallTime:
      type: string
      description: |
        24-hour local wall-clock time. Stored alongside an IANA `tz`; the
        instant is a materialisation of the next occurrence, never the source
        of truth. Storing only the instant is the fixed-offset bug in a
        different costume.
      pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
      examples: ["07:00"]
    Timezone:
      type: string
      description: IANA zone name. Never a fixed UTC offset.
      examples: ["America/Chicago"]
    OkEnvelope:
      type: object
      required: [success]
      properties:
        success: { const: true }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: "#/components/schemas/ApiError" }
    ErrorCode:
      type: string
      description: |
        The typed code is the contract; the HTTP status is advisory. Clients
        branch on this, never on the status line.

        `PAIRING_CODE_EXPIRED` and `PAIRING_CODE_ALREADY_USED` are the only
        places the service tells one authentication failure from another, and
        they are safe because **both are reachable only by presenting a code
        that matches a stored one**. A wrong guess, and an attempt made while
        no code is live, are the same `UNAUTHORIZED`. So the extra detail
        reaches the person holding the code and nobody else — which is what
        lets a pairing screen on a phone say something more useful than "that
        did not work".
      enum:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - PAIRING_CODE_EXPIRED
        - PAIRING_CODE_ALREADY_USED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - IDEMPOTENCY_KEY_REUSE
        - IDEMPOTENCY_KEY_REQUIRED
        - DEFERRAL_NOT_LATER
        - RRULE_UNSUPPORTED
        - UNKNOWN_JOB_KIND
        - DEVICE_TOKEN_INVALID
        - QUIET_HOURS
        - RATE_LIMITED
        - UPSTREAM_UNAVAILABLE
        - INTERNAL
    ApiError:
      type: object
      required: [code, message, retryable]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message:
          type: string
          description: Human-readable, for a log line or an admin screen. Never parsed.
        retryable:
          type: boolean
          description: |
            Whether the client's backoff should try again. `RATE_LIMITED`,
            `UPSTREAM_UNAVAILABLE` and `INTERNAL` are retryable; a validation
            failure never is. Auth and billing failures halt rather than
            retry — retrying an auth failure fifty times is the worst
            possible response to a key having shadowed the subscription login.
        details:
          type: [object, "null"]
          description: Structured context. `VALIDATION_FAILED` carries `{ field, reason }[]`.
          additionalProperties: true
        retryAfterMs:
          type: [integer, "null"]
          description: Server-suggested backoff floor, when it knows one.
...
    ReminderKind:
      type: string
      description: |
        Drives the catch-up policy, which is why it is stored on the row.

        A `commitment` never collapses: it fires late and says it is late.
        A `rhythm` message does the opposite — yesterday's morning agenda is
        worthless today, so `supersedesPrevious` is set and only the current
        one fires, with one line saying what was skipped so even the
        suppression is visible.
      enum: [commitment, rhythm]
    ReminderOrigin:
      type: string
      enum: [he_asked, she_noticed]
      description: |
        `he_asked` — he requested it in the conversation.
        `she_noticed` — she offered it unprompted.

        DERIVED rather than claimed wherever it can be: a heartbeat or dream
        turn has no message from him at all, so "she thought of it" is a fact
        about that turn rather than a self-report — and those are exactly the
        3am ones. Same rule as `urgentBecauseHeSaid`: a conclusion can only be
        trusted, evidence can be checked.
    ReminderDeliveryState:
      type: string
      description: |
        `delivered` means APNs accepted the push. `acknowledged` means the
        device confirmed it. Only the second one satisfies the guarantee.
      enum:
        - scheduled
        - due
        - delivered
        - acknowledged
        - deferred
        - completed
        - cancelled
        - failed
    Reminder:
      type: object
      required:
        [
          id,
          kind,
          text,
          todoId,
          eventId,
          wallTime,
          tz,
          rrule,
          scheduledFor,
          nextFireAt,
          urgent,
          late,
          deferredFrom,
          supersedesPrevious,
          deliveryState,
          createdAt,
          updatedAt,
          completedAt,
        ]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        kind: { $ref: "#/components/schemas/ReminderKind" }
        text:
          type: string
          description: |
            Composed at creation time, in Syl's voice. Delivery is a zero-turn
            path and reads this verbatim — no model is in the way to improve
            it later.
        because:
          type: [string, "null"]
          description: |
            Why this reminder exists, in his words or hers. `syl-y82`.

            `remind_me` has REQUIRED this since it shipped and discarded it
            until now — on the one verb that wakes him. `SOUL.md` promises he
            can tell a good suggestion from a wrong one and can tell her to
            stop making a kind he dislikes; both were false while it was
            dropped.

            Null on reminders written before the column existed. Not
            backfilled: a guessed reason is exactly the claim-beyond-evidence
            this field exists to prevent.
        origin:
          oneOf:
            - { $ref: "#/components/schemas/ReminderOrigin" }
            - { type: "null" }
          description: |
            Whether HE asked for it or SHE thought of it — a different question
            from `because`, and the one a careful reader actually got wrong.
            Prose answers "why does this exist" and only sometimes "did he
            ask". A list showing him which are hers needs this, not that.
        todoId:
          oneOf:
            - { $ref: "#/components/schemas/Id" }
            - { type: "null" }
        eventId:
          oneOf:
            - { $ref: "#/components/schemas/Id" }
            - { type: "null" }
        wallTime: { $ref: "#/components/schemas/WallTime" }
        tz: { $ref: "#/components/schemas/Timezone" }
        rrule:
          type: [string, "null"]
          description: |
            RFC 5545 RRULE text, restricted to a deliberate subset: daily,
            weekly-by-day, monthly-by-date, yearly. Anything outside it is
            rejected at write time with `RRULE_UNSUPPORTED` rather than
            half-supported — a rule that parses but schedules wrongly is a
            silently dropped reminder.
          examples: ["FREQ=WEEKLY;BYDAY=MO,WE,FR"]
        scheduledFor:
          allOf: [{ $ref: "#/components/schemas/Instant" }]
          description: |
            The instant this occurrence was originally due. Survives deferral
            and lateness so the client can render "this was due Tuesday at
            09:00".
        nextFireAt:
          allOf: [{ $ref: "#/components/schemas/Instant" }]
          description: The materialised instant of the next occurrence.
        urgent:
          type: boolean
          description: |
            Maps one-to-one onto APNs `interruption-level: time-sensitive`,
            and onto the `alert` presence state. Rationed by being coupled to
            something already scarce.
        late:
          type: boolean
          description: |
            Fired after `scheduledFor` because the machine was down. A late
            reminder says it is late; it is never silently dropped and never
            pretends to be on time.
        deferredFrom:
          oneOf:
            - { $ref: "#/components/schemas/Instant" }
            - { type: "null" }
          description: The previous `nextFireAt` in the deferral chain.
        supersedesPrevious:
          type: boolean
          description: Rhythm reminders only. Suppressed occurrences are reported, not hidden.
        deliveryState: { $ref: "#/components/schemas/ReminderDeliveryState" }
        createdAt: { $ref: "#/components/schemas/Instant" }
        updatedAt: { $ref: "#/components/schemas/Instant" }
        completedAt:
          oneOf:
            - { $ref: "#/components/schemas/Instant" }
            - { type: "null" }
    ReminderPage:
      type: object
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Reminder" }
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }
    CreateReminderRequest:
      type: object
      required: [text, wallTime, tz]
      properties:
        text: { type: string, minLength: 1 }
        kind: { $ref: "#/components/schemas/ReminderKind" }
        wallTime: { $ref: "#/components/schemas/WallTime" }
        tz: { $ref: "#/components/schemas/Timezone" }
        date:
          type: [string, "null"]
          description: "`YYYY-MM-DD` local date for a one-shot. Null with an `rrule`."
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        rrule: { type: [string, "null"] }
        todoId:
          oneOf:
            - { $ref: "#/components/schemas/Id" }
            - { type: "null" }
        urgent: { type: boolean, default: false }
    UpdateReminderRequest:
      type: object
      description: Every field optional; omitted fields are left alone.
      properties:
        text: { type: string, minLength: 1 }
        wallTime: { $ref: "#/components/schemas/WallTime" }
        tz: { $ref: "#/components/schemas/Timezone" }
        rrule: { type: [string, "null"] }
        urgent: { type: boolean }
    SnoozeReminderRequest:
      type: object
      description: |
        Supply exactly one of `until` or `minutes`. Whichever is used, the
        resulting instant must be strictly later than the current
        `nextFireAt`, or the call fails with `DEFERRAL_NOT_LATER`.
      properties:
        until:
          oneOf:
            - { $ref: "#/components/schemas/Instant" }
            - { type: "null" }
        minutes:
          type: [integer, "null"]
          minimum: 1
          examples: [15]

    # ----------------------------------------------------------- todos ---
...